Split state out into seperate objects
