+++ lib/enter_weight_page.dart
@@ -1,3 +1,4 @@
+import 'package:flexify/app_state.dart';
 import 'package:flexify/database.dart';
 import 'package:flexify/main.dart';
 import 'package:flutter/material.dart';
